Packaging C API in a Capsule, following Python docs

+++ src/pocky/ext/include/pocky_bufpair.h
@@ -34,7 +34,7 @@ extern PyMethodDef pocky_bufpair_methods[];
 extern int pocky_bufpair_empty_like(pocky_context_object *context,
     pocky_bufpair_object *like, pocky_bufpair_object **bufpair);
 extern int pocky_bufpair_empty_from_shape(pocky_context_object *context,
-    size_t ndim, size_t *shape, pocky_bufpair_object **bufpair);
+    size_t ndim, long *shape, pocky_bufpair_object **bufpair);
 
 #endif      /* POCKY_BUFPAIR_H */
